feat: 리뷰 수와 평점 계산

+++ src/main/java/life/walkit/server/trail/dto/response/TrailListResponse.java
@@ -11,10 +11,10 @@ public record TrailListResponse(
     String location,
     Double length,
     String routeImageUrl,
-    Integer reviewCount,
+    Long reviewCount,
     Double rating
 ) {
-    public static TrailListResponse from(Trail trail, Optional<TrailWalkImage> trailWalkImageOpt, int reviewCount, double rating) {
+    public static TrailListResponse from(Trail trail, Optional<TrailWalkImage> trailWalkImageOpt, long reviewCount, double rating) {
         String imageUrl = trailWalkImageOpt.map(TrailWalkImage::getRouteImage).orElse(null);
 
         return new TrailListResponse(
